Json duplicate properties/ array vs object (#17)
* add environment name to deploy job
do direct checks on properties between top and data and detect name clash

* handle array vs single object

* small tweak

* name int test jobs per env

* Update JsonElementExtensionTests.cs

* Refactor JSON handling and update tests

Refactored JSON handling in `SapActBaseWorker.cs` to use a switch expression for root elements and replaced the loop with a `foreach` loop. Updated telemetry metric tracking for better handling of single and multiple items. Modified `GenerateColumnList` in `JsonElementExtensions.cs` to initialize with an empty list. Refactored `ExportToFlattenedDictionary` to add top-level fields after data fields and removed redundant key checks. Corrected JSON indentation in `JsonElementExtensionTests.cs`.

* Convert JsonElement enumeration to List in Array handling

The code change modifies the handling of `JsonValueKind.Array` in the
switch statement. Previously, `jsonDocument.RootElement.EnumerateArray()`
returned an `IEnumerable<JsonElement>`. The updated code converts this
`IEnumerable<JsonElement>` to a `List<JsonElement>` by calling `.ToList()`.

This change ensures that when the `jsonDocument.RootElement` is of type
`JsonValueKind.Array`, the elements are enumerated and then converted to
a `List<JsonElement>`. This facilitates operations that require a list
rather than an enumerable, such as indexing or certain collection
manipulations.

* refactor column list function
remove unneeded use in SQL service

* Simplify SelectMany call in GenerateColumnList method

The SelectMany method call in the GenerateColumnList method of the JsonElementExtensions class has been simplified. The explicit type parameters have been removed, and a call to ToArray() has been added for the EnumerateObject() result when the property name is "data".